fix(validate): make the tenure check a warning, and stop it being an off switch
The calibration now asserts on the EXTRACTED column, not only the flagged one.
That gap is why both of this round's over-corrections shipped past a green
harness: a check that has been switched off and a check that has been fixed
both report zero false positives. The set of fixtures whose extraction is
non-empty is pinned as a hand-written membership list — a set compared against
itself cannot notice a deletion, which is the exact failure being guarded.

The unreadable-tenure guard was an off switch. It matched any letter-word
before any year-word anywhere in the source, so one bullet reading "Cut cloud
spend by 1.2M USD per year" silenced factual.inflated_experience for the whole
document; "year over year growth", "the fiscal year close" and "last year" did
the same. It is now scoped by the admission rule the claims side uses, so
"several years of experience" still silences and an impact figure does not.

Removing the summary bypass cost six of twenty-one fixtures their extraction,
not the one previously reported — including the two sentences this repo's own
truthful fixtures write verbatim ("Eight years of backend work, most of it on
payment systems", "Backend engineer with 8 years building distributed
services"). Removing it was right; removing it without widening what the gate
admits was not. The summary admission is back and shape-aware: the span must
open a clause, preceded by nothing or by a lead-in word, and be followed by a
preposition or participle. The three legacy-system sentences fail both halves
independently.

Attributive vendor terms are no longer credentials. "Docker Certified images",
"Red Hat certified build images", "VMware certified storage arrays" and
"Certified Scrum team" are grammatically identical to a real credential, and
adjacency alone could not tell them apart — seven of seven fired. A
certification names a ROLE its holder is certified to fill, so one must appear
within a bounded window after it.

Decades were covered by neither the number table nor the guard, and both a
comment and a test doc claimed otherwise. "over a decade of experience" now
reads as an unmeasurable tenure rather than an absent one; it is not mapped to
ten, because "over a decade" is anywhere from ten to nineteen and picking a
number invents evidence on the sparing side.

A numeric month after an opener no longer closes a span. "Seit 03/2016" wants a
year within one word of the opener to satisfy is_open_ended, so the current role
read as closed at its own start year.

Two MEDIUMs: one certification named in both its forms is one finding, not two;
and the "Source is a superset of Claims" claim is now stated with the bound it
actually holds under.

## The severity call

factual.inflated_experience ships as a WARNING. The criterion permitted a
Critical — zero false positives across 25 truthful documents while reading the
tenure sentences this repo's fixtures write verbatim — and the argument against
is the shape of the input rather than the number. Two review rounds found five
distinct registers that each produced a false Critical, and the corpus was green
before every one of them was named; the claims side reads unbounded prose in
seven languages discriminated by three hand-curated word lists. The cost is
asymmetric too: this code routes to SectionKey::Summary, so a false one would
spend provider calls rewriting a correct summary and could pressure the model
into understating a tenure the candidate really has. A Warning cannot —
criticals_by_section never sees it, and it leaves FABRICATION_CODES with the
other Warning.

factual.unsourced_certification stays Critical: its trigger is three bounded
vocabularies intersecting, not prose.

credentials.rs reached 1397 of R8's 1400-line hard cap, so the three families
move into credentials/{tenure,certifications,education}.rs behind a flat
re-export. The parent keeps the dispatcher and the one helper both certification
and education passes share.
